add protobuf tests for spring-kafka to make sure that ser/de abstraction works when needed
