+++ lib/stove-testing-e2e/build.gradle.kts
@@ -8,7 +8,7 @@ dependencies {
   api(libs.jackson.kotlin)
   api(libs.jackson.arrow)
   api(libs.google.gson)
-  api(libs.kotlinx.serialization)
+  api(libs.kotlinx.serialization.json)
   api(libs.testcontainers) {
     version {
       require(libs.testcontainers.asProvider().get().version!!)
