+++ starters/spring/stove-spring-testing-e2e-kafka/src/test/kotlin/com/trendyol/stove/testing/e2e/kafka/shared.kt
@@ -7,3 +7,5 @@ class Setup : AbstractProjectConfig() {
   @ExperimentalKotest
   override val concurrentSpecs: Int = 1
 }
+
+class StoveBusinessException(message: String) : Exception(message)
